Quick port over to Couchbase
